requirements.txt
beautifulsoup4==4.9.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
html5lib==1.0.1
idna==2.8
pycparser==2.19
PyJWT==1.7.1
python-telegram-bot==12.4.2
requests==2.22.0
six==1.13.0
soupsieve==2.0
tornado==6.0.3
urllib3==1.25.7
webencodings==0.5.1
